FAST_API/requirements.txt
SQLAlchemy
flask-sqlalchemy
httpx
fastapi
black
flake8
flake8-bugbear
flake8-pie
isort
mypy
aiosqlite
pytest-asyncio
pydantic